move "remove unused voice" from Beat to Measure menu
