+++ common/resources/lang/messages.properties
@@ -396,12 +396,12 @@ measure.clean=Clean Measure
 measure.remove=Remove Measure
 measure.copy=Copy Measure
 measure.paste=Paste Measure
+measure.voice.remove-unused=Remove unused Voice
 
 beat=Beat
 beat.insert-rest=Insert Rest Beat
 beat.delete-note-or-rest=Delete Note or Rest
 beat.clean=Clean Beat
-beat.voice.remove-unused=Remove unused Voice
 text.insert=Insert Text
 text.editor=Text Editor
 text.text=Text
@@ -786,7 +786,7 @@ action.measure.go-first=Go to first Measure
 action.measure.go-last=Go to last Measure
 action.measure.go-next=Go to next Measure
 action.measure.go-previous=Go to previous Measure
-action.beat.general.remove-unused-voice=Remove unused Voice
+action.measure.general.remove-unused-voice=Remove unused Voice
 action.beat.general.voice-up=Set Stems up
 action.beat.general.voice-down=Set Stems down
 action.beat.general.voice-auto=Set Auto Stems
