+++ common/resources/lang/messages_bg.properties
@@ -252,12 +252,12 @@ measure.clean=Изчистване на такт
 measure.remove=Премахване на такт
 measure.copy=Копиране на такт
 measure.paste=Поставяне на такт
+measure.voice.remove-unused=Премахване на неизползван глас
 
 beat=Удар
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Изчистване на удар
-beat.voice.remove-unused=Премахване на неизползван глас
 text.insert=Вмъкване на текст
 text.editor=Редактор на текст
 text.text=Текст
@@ -642,7 +642,7 @@ action.measure.go-first=Отиване до първия такт
 action.measure.go-last=Отиване до последния такт
 action.measure.go-next=Отиване до следващия такт
 action.measure.go-previous=Отиване до предишен такт
-action.beat.general.remove-unused-voice=Премахване на неизползван глас
+action.measure.general.remove-unused-voice=Премахване на неизползван глас
 # action.beat.general.voice-up=
 # action.beat.general.voice-down=
 # action.beat.general.voice-auto=
