+++ common/resources/lang/messages_ca.properties
@@ -252,12 +252,12 @@ measure.clean=Neteja compàs
 measure.remove=Esborra compàs
 measure.copy=Copia compàs
 measure.paste=Enganxa compàs
+measure.voice.remove-unused=Esborra veu no usada
 
 # beat=
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Neteja beat
-beat.voice.remove-unused=Esborra veu no usada
 text.insert=Insereix text
 text.editor=Editor de text
 # text.text=
@@ -642,7 +642,7 @@ action.measure.go-first=Vés al primer compàs
 action.measure.go-last=Vés al darrer compàs
 action.measure.go-next=Vés al compàs següent
 action.measure.go-previous=Vés al compàs anterior
-action.beat.general.remove-unused-voice=Esborra veu no usada
+action.measure.general.remove-unused-voice=Esborra veu no usada
 action.beat.general.voice-up=Assigna stems amunt
 action.beat.general.voice-down=Assigna stems avall
 action.beat.general.voice-auto=Assigna stems automàtics
