+++ common/resources/lang/messages_cs.properties
@@ -252,12 +252,12 @@ measure.clean=Vymazat obsah taktu
 measure.remove=Odebrat takt
 measure.copy=Kopírovat takt
 measure.paste=Vložit takt
+# measure.voice.remove-unused=
 
 beat=Nota
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Čistá doba
-# beat.voice.remove-unused=
 # text.insert=
 # text.editor=
 # text.text=
@@ -642,7 +642,7 @@ action.measure.go-first=Jít na první takt
 action.measure.go-last=Jít na poslední takt
 action.measure.go-next=Jít na následnující takt
 action.measure.go-previous=Jít na předchozí takt
-# action.beat.general.remove-unused-voice=
+# action.measure.general.remove-unused-voice=
 # action.beat.general.voice-up=
 # action.beat.general.voice-down=
 # action.beat.general.voice-auto=
