+++ common/resources/lang/messages_de.properties
@@ -252,12 +252,12 @@ measure.clean=Takt leeren
 measure.remove=Takt löschen
 measure.copy=Takt kopieren
 measure.paste=Takt einfügen
+measure.voice.remove-unused=Ungenutzte Stimme entfernen
 
 beat=Taktschlag
 beat.insert-rest=Pause hinzufügen
 beat.delete-note-or-rest=Note/Pause löschen
 beat.clean=Schlag leeren
-beat.voice.remove-unused=Ungenutzte Stimme entfernen
 text.insert=Text hinzufügen
 text.editor=Texteditor
 # text.text=
@@ -642,7 +642,7 @@ action.measure.go-first=Zum ersten Takt springen
 action.measure.go-last=Zum letzten Takt springen
 action.measure.go-next=Zum nächsten Takt springen
 action.measure.go-previous=Zum vorherigen Takt springen
-action.beat.general.remove-unused-voice=Ungenutzte Stimme entfernen
+action.measure.general.remove-unused-voice=Ungenutzte Stimme entfernen
 action.beat.general.voice-up=Notenhälse nach oben ausrichten
 action.beat.general.voice-down=Notenhälse nach unten ausrichten
 action.beat.general.voice-auto=Notenhälse automatisch ausrichten
