+++ common/resources/lang/messages_el.properties
@@ -252,12 +252,12 @@ measure.clean=Καθαρισμός Μέτρου
 measure.remove=Αφαίρεση Μέτρου
 measure.copy=Αντιγραφή Μέτρου
 measure.paste=Προορισμός Μέτρου
+measure.voice.remove-unused=Αφαίρεση Μη Χρησιμοποιούμενης Φωνής
 
 # beat=
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Καθαρισμός Beat
-beat.voice.remove-unused=Αφαίρεση Μη Χρησιμοποιούμενης Φωνής
 text.insert=Εισαγωγή Κειμένου
 text.editor=Επεξεργαστής Κειμένου
 text.text=Κείμενο
@@ -642,7 +642,7 @@ action.measure.go-first=Πρόσθεση στο πρώτο μέτρο
 action.measure.go-last=Πρόσθεση στο τελευταίο μέτρο
 action.measure.go-next=Πρόσθεση από στο επόμενο μέτρο
 action.measure.go-previous=Πρόσθεση από στο προηγούμενο μέτρο
-action.beat.general.remove-unused-voice=Αφαίρεση Μη Χρησιμοποιούμενης Φωνής
+action.measure.general.remove-unused-voice=Αφαίρεση Μη Χρησιμοποιούμενης Φωνής
 action.beat.general.voice-up=Όρισε τα άνω στελέχη
 action.beat.general.voice-down=Όρισε τα κάτω στελέχη
 action.beat.general.voice-auto=Όρισε αυτόματα τα στελέχη
