+++ common/resources/lang/messages_es.properties
@@ -252,12 +252,12 @@ measure.clean=Limpiar compás
 measure.remove=Borrar compás
 measure.copy=Copiar compás
 measure.paste=Pegar compás
+measure.voice.remove-unused=Eliminar la voz no utilizada
 
 beat=Nota
 beat.insert-rest=Insertar Silencio
 beat.delete-note-or-rest=Borrar Nota/Silencio
 beat.clean=Limpiar Notas
-beat.voice.remove-unused=Eliminar la voz no utilizada
 text.insert=Insertar Texto
 text.editor=Editor de Texto
 text.text=Texto
@@ -642,7 +642,7 @@ action.measure.go-first=Ir al primer compás
 action.measure.go-last=Ir al último compás
 action.measure.go-next=Ir al siguiente compás
 action.measure.go-previous=Ir al compás anterior
-action.beat.general.remove-unused-voice=Eliminar la voz no utilizada
+action.measure.general.remove-unused-voice=Eliminar la voz no utilizada
 action.beat.general.voice-up=Notas hacia arriba
 action.beat.general.voice-down=Notas hacia abajo
 action.beat.general.voice-auto=Dirección de notas automatica
