+++ common/resources/lang/messages_eu.properties
@@ -252,12 +252,12 @@ measure.clean=Ezabatu Kompasa
 measure.remove=Kendu kompasa
 measure.copy=Kopiatu Kompasa
 measure.paste=itsatsi Kompasa
+# measure.voice.remove-unused=
 
 beat=Neurria
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Ezabatu neurria
-# beat.voice.remove-unused=
 text.insert=Textua idatzi
 text.editor=Textu editorea
 text.text=Textua
@@ -642,7 +642,7 @@ action.measure.go-first=Lehengo kompasaraino
 action.measure.go-last=Azken kompasaraino
 action.measure.go-next=Hurrengo kompasaraino
 action.measure.go-previous=Aurreko kompasera
-# action.beat.general.remove-unused-voice=
+# action.measure.general.remove-unused-voice=
 # action.beat.general.voice-up=
 # action.beat.general.voice-down=
 # action.beat.general.voice-auto=
