+++ common/resources/lang/messages_fi.properties
@@ -252,12 +252,12 @@ measure.clean=Tyhjennä tahti
 measure.remove=Poista tahti
 measure.copy=Kopioi tahti
 measure.paste=Liitä tahti
+# measure.voice.remove-unused=
 
 beat=Isku
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Tyhjennä isku
-# beat.voice.remove-unused=
 text.insert=Lisää tekstiä
 text.editor=Tekstieditori
 text.text=Teksti
@@ -642,7 +642,7 @@ action.measure.go-first=Siirry ensimmäiseen tahtiin
 action.measure.go-last=Siirry viimeiseen tahtiin
 action.measure.go-next=Siirry seuraavaan tahtiin
 action.measure.go-previous=Siirry edelliseen tahtiin
-action.beat.general.remove-unused-voice=Poista käyttämättömät äänet
+action.measure.general.remove-unused-voice=Poista käyttämättömät äänet
 action.beat.general.voice-up=Siirrä varret ylös
 action.beat.general.voice-down=Siirrä varret alas
 action.beat.general.voice-auto=Aseta varret automaattisesti
