+++ common/resources/lang/messages_fr.properties
@@ -252,12 +252,12 @@ measure.clean=Effacer la mesure
 measure.remove=Supprimer la mesure
 measure.copy=Copier la mesure
 measure.paste=Coller la mesure
+measure.voice.remove-unused=Supprimer la voix inutilisée
 
 beat=Note
 beat.insert-rest=Insérer un silence
 beat.delete-note-or-rest=Supprimer la note ou le silence
 beat.clean=Effacer la note
-beat.voice.remove-unused=Supprimer la voix inutilisée
 text.insert=Insérer un texte
 text.editor=Éditeur de texte
 text.text=Texte
@@ -642,7 +642,7 @@ action.measure.go-first=Aller à la première mesure
 action.measure.go-last=Aller à la dernière mesure
 action.measure.go-next=Aller à la mesure suivante
 action.measure.go-previous=Aller à la mesure précédente
-action.beat.general.remove-unused-voice=Supprimer la voix inutilisée
+action.measure.general.remove-unused-voice=Supprimer la voix inutilisée
 action.beat.general.voice-up=Définir une hampe haute
 action.beat.general.voice-down=Définir une hampe basse
 action.beat.general.voice-auto=Définir les hampes automatiques
