+++ common/resources/lang/messages_hu.properties
@@ -252,12 +252,12 @@ measure.clean=Csendes ütem
 measure.remove=Ütem eltávolítása
 measure.copy=Ütem másolása
 measure.paste=Ütem beillesztése
+# measure.voice.remove-unused=
 
 beat=Hangjegy
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Hang/akkord törlése
-# beat.voice.remove-unused=
 # text.insert=
 # text.editor=
 # text.text=
@@ -642,7 +642,7 @@ action.measure.go-first=Ugrás az elsõ ütemre
 action.measure.go-last=Ugrás az utolsó ütemre
 action.measure.go-next=Ugrás a következõ ütemre
 action.measure.go-previous=Ugrás az elõzõ ütemre
-# action.beat.general.remove-unused-voice=
+# action.measure.general.remove-unused-voice=
 # action.beat.general.voice-up=
 # action.beat.general.voice-down=
 # action.beat.general.voice-auto=
