+++ common/resources/lang/messages_it.properties
@@ -252,12 +252,12 @@ measure.clean=Svuota misura
 measure.remove=Rimuovi misura
 measure.copy=Copia misura
 measure.paste=Incolla misura
+measure.voice.remove-unused=Elimina voce non utilizzata
 
 beat=Battuta
 beat.insert-rest=Inserisci pausa
 beat.delete-note-or-rest=Elimina nota o pausa
 beat.clean=Svuota battuta
-beat.voice.remove-unused=Elimina voce non utilizzata
 text.insert=Inserisci testo
 text.editor=Editor di testo
 text.text=Testo
@@ -642,7 +642,7 @@ action.measure.go-first=Vai alla prima misura
 action.measure.go-last=Vai all'ultima misura
 action.measure.go-next=Vai alla prossima misura
 action.measure.go-previous=Vai alla misura precedente
-action.beat.general.remove-unused-voice=Elimina voce non utilizzata
+action.measure.general.remove-unused-voice=Elimina voce non utilizzata
 action.beat.general.voice-up=Gambi verso l'alto
 action.beat.general.voice-down=Gambi verso il basso
 action.beat.general.voice-auto=Orientamento automatico dei gambi
