+++ common/resources/lang/messages_ja.properties
@@ -252,12 +252,12 @@ measure.clean=クリア (選択小節)
 measure.remove=削除 (範囲指定)
 measure.copy=コピー (範囲指定)
 measure.paste=貼り付け
+measure.voice.remove-unused=未使用音声の除去
 
 beat=音符/エフェクト
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=1拍除去
-beat.voice.remove-unused=未使用音声の除去
 text.insert=テキスト挿入
 text.editor=テキスト・エディタ
 text.text=テキスト
@@ -642,7 +642,7 @@ action.measure.go-first=小節: 先頭へ
 action.measure.go-last=小節: 最後へ
 action.measure.go-next=小節: 次へ
 action.measure.go-previous=小節: 前へ
-action.beat.general.remove-unused-voice=音符/エフェクト: 未使用音声の除去
+action.measure.general.remove-unused-voice=音符/エフェクト: 未使用音声の除去
 action.beat.general.voice-up=音符/エフェクト: 符尾の向きを上にする
 action.beat.general.voice-down=音符/エフェクト: 符尾の向きを下にする
 action.beat.general.voice-auto=音符/エフェクト: 符尾の向きの自動調整
