+++ common/resources/lang/messages_ko.properties
@@ -252,12 +252,12 @@ measure.clean=마디 비우기
 measure.remove=마디 삭제하기
 measure.copy=마디 복사
 measure.paste=마디 붙여넣기
+# measure.voice.remove-unused=
 
 beat=박자
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=박자 비우기
-# beat.voice.remove-unused=
 text.insert=텍스트 삽입
 text.editor=텍스트 편집기
 # text.text=
@@ -642,7 +642,7 @@ action.measure.go-first=첫 마디로 이동
 action.measure.go-last=마지막 마디로 이동
 action.measure.go-next=다음 마디로 이동
 action.measure.go-previous=이전 마디로 이동
-# action.beat.general.remove-unused-voice=
+# action.measure.general.remove-unused-voice=
 # action.beat.general.voice-up=
 # action.beat.general.voice-down=
 # action.beat.general.voice-auto=
