+++ common/resources/lang/messages_lt.properties
@@ -252,12 +252,12 @@ measure.clean=Nuimti takto ženklą
 measure.remove=Pašalinti taktą
 measure.copy=Kopijuoti taktą
 measure.paste=Įdėti taktą
+measure.voice.remove-unused=Nenaudojamo balso pašalinimas
 
 beat=Natos
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Akcentavimo nuėmimas
-beat.voice.remove-unused=Nenaudojamo balso pašalinimas
 text.insert=Įterpti tekstą
 text.editor=Teksto redaktorius
 text.text=Tekstas
@@ -642,7 +642,7 @@ action.measure.go-first=Eiti į pirmą taktą
 action.measure.go-last=Eiti į paskutinį taktą
 action.measure.go-next=Eiti į kitą taktą
 action.measure.go-previous=Eiti į ankstesnį taktą
-action.beat.general.remove-unused-voice=Nenaudojamo balso pašalinimas
+action.measure.general.remove-unused-voice=Nenaudojamo balso pašalinimas
 action.beat.general.voice-up=Penklinę papildyti aukštyn
 action.beat.general.voice-down=Penklinę papildyti žemyn
 action.beat.general.voice-auto=Automatinis penklinės papildymas linijomis
