+++ common/resources/lang/messages_nl.properties
@@ -252,12 +252,12 @@ measure.clean=Wis maat
 measure.remove=Verwijder maat
 measure.copy=Kopieer maat
 measure.paste=Plak maat
+# measure.voice.remove-unused=
 
 # beat=
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Wis Beat
-# beat.voice.remove-unused=
 text.insert=Tekst Invoegen
 text.editor=Tekst Editor
 text.text=Tekst
@@ -642,7 +642,7 @@ action.measure.go-first=Ga naar eerst maat
 action.measure.go-last=Ga naar laatste maat
 action.measure.go-next=Ga naar volgende maat
 action.measure.go-previous=Ga naar vorige maat
-# action.beat.general.remove-unused-voice=
+# action.measure.general.remove-unused-voice=
 # action.beat.general.voice-up=
 # action.beat.general.voice-down=
 # action.beat.general.voice-auto=
