+++ common/resources/lang/messages_pl.properties
@@ -252,12 +252,12 @@ measure.clean=Wyczyść takt
 measure.remove=Usuń takt
 measure.copy=Kopiuj takt
 measure.paste=Wstaw takt
+measure.voice.remove-unused=Usuń nieużywany głos
 
 beat=Nuta
 beat.insert-rest=Wstaw pauzę
 beat.delete-note-or-rest=Usuń nutę lub pauzę
 beat.clean=Wyczyść wartość
-beat.voice.remove-unused=Usuń nieużywany głos
 text.insert=Wstaw tekst
 text.editor=Edytor tekstu
 text.text=Tekst
@@ -642,7 +642,7 @@ action.measure.go-first=Idź do pierwszego taktu
 action.measure.go-last=Idź do ostatniego taktu
 action.measure.go-next=Idź do następnego taktu
 action.measure.go-previous=Idź do poprzedniego taktu
-action.beat.general.remove-unused-voice=Usuń nieużywany głos
+action.measure.general.remove-unused-voice=Usuń nieużywany głos
 action.beat.general.voice-up=Ustaw laseczki w górę
 action.beat.general.voice-down=Ustaw laseczki w dół
 action.beat.general.voice-auto=Ustaw automatyczny kierunek laseczek
