+++ common/resources/lang/messages_pt.properties
@@ -252,12 +252,12 @@ measure.clean=Limpar Compasso
 measure.remove=Remover Compasso
 measure.copy=Copiar Compasso
 measure.paste=Colar Compasso
+measure.voice.remove-unused=Remover Voz não utilizada
 
 beat=Nota
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Limpar batida
-beat.voice.remove-unused=Remover Voz não utilizada
 text.insert=Inserir Texto
 # text.editor=
 # text.text=
@@ -642,7 +642,7 @@ action.measure.go-first=Ir para o primeiro compasso
 action.measure.go-last=Ir para o último compasso
 action.measure.go-next=Ir para o próximo compasso
 action.measure.go-previous=Ir para o compasso anterior
-action.beat.general.remove-unused-voice=Remover voz não usada
+action.measure.general.remove-unused-voice=Remover voz não usada
 action.beat.general.voice-up=Configura bastões para cima
 action.beat.general.voice-down=Configura bastões para baixo
 action.beat.general.voice-auto=Configura bastões automáticos
