+++ common/resources/lang/messages_ru.properties
@@ -250,12 +250,12 @@ measure.clean=Очистить
 measure.remove=Удалить
 measure.copy=Копировать
 measure.paste=Вставить
+measure.voice.remove-unused=Удалить неиспользуемый голос
 
 beat=Доля
 beat.insert-rest=Вставить долю паузы
 beat.delete-note-or-rest=Удалить ноту или паузу
 beat.clean=Очистить
-beat.voice.remove-unused=Удалить неиспользуемый голос
 text.insert=Вставить текст
 text.editor=Редактор текста
 text.text=Текст
@@ -640,7 +640,7 @@ action.measure.go-first=Перейти к первый такту
 action.measure.go-last=Перейти к последнему такту
 action.measure.go-next=Перейти к следующему такту
 action.measure.go-previous=Перейти к предыдущему такту
-action.beat.general.remove-unused-voice=Удалить неиспользуемый голос
+action.measure.general.remove-unused-voice=Удалить неиспользуемый голос
 action.beat.general.voice-up=Штили вверх
 action.beat.general.voice-down=Штили вниз
 action.beat.general.voice-auto=Авторасстановка штилей
