+++ common/resources/lang/messages_sr.properties
@@ -252,12 +252,12 @@ measure.clean=Obriši sve iz takta
 measure.remove=Obriši takt
 measure.copy=Kopiraj takt
 measure.paste=Zalepi takt
+# measure.voice.remove-unused=
 
 beat=Nota
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Očisti notu
-# beat.voice.remove-unused=
 text.insert=Ubaci tekst
 # text.editor=
 # text.text=
@@ -642,7 +642,7 @@ action.measure.go-first=Idi na prvi takt
 action.measure.go-last=Idi na poslednji takt
 action.measure.go-next=Idi na sledeći takt
 action.measure.go-previous=Idi na prethodni takt
-# action.beat.general.remove-unused-voice=
+# action.measure.general.remove-unused-voice=
 # action.beat.general.voice-up=
 # action.beat.general.voice-down=
 # action.beat.general.voice-auto=
