+++ common/resources/lang/messages_sv.properties
@@ -252,12 +252,12 @@ measure.clean=Rensa takt
 measure.remove=Ta bort takt
 measure.copy=Kopiera takt
 measure.paste=Klistra in takt
+measure.voice.remove-unused=Ta bort oanvänd stämma
 
 beat=Slag
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Rensa slag
-beat.voice.remove-unused=Ta bort oanvänd stämma
 text.insert=Infoga text
 text.editor=Textredigerare
 # text.text=
@@ -642,7 +642,7 @@ action.measure.go-first=Gå till första takten
 action.measure.go-last=Gå till sista takten
 action.measure.go-next=Gå till nästa takt
 action.measure.go-previous=Gå till föregående takt
-action.beat.general.remove-unused-voice=Ta bort oanvänd stämma
+action.measure.general.remove-unused-voice=Ta bort oanvänd stämma
 action.beat.general.voice-up=Sätt notskaft uppåt
 action.beat.general.voice-down=Sätt notskaft nedåt
 action.beat.general.voice-auto=Sätt notskaft automatiskt
