+++ common/resources/lang/messages_uk.properties
@@ -252,12 +252,12 @@ measure.clean=Очистити такт
 measure.remove=Видалити такт
 measure.copy=Копіювати такт
 measure.paste=Вставити такт
+# measure.voice.remove-unused=
 
 beat=Доля
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Очистити долю
-# beat.voice.remove-unused=
 text.insert=Вставити текст
 text.editor=Редагування тексту
 text.text=Текст
@@ -642,7 +642,7 @@ action.measure.go-first=Перший такт
 action.measure.go-last=Останній такт
 action.measure.go-next=Наступний такт
 action.measure.go-previous=Попередній такт
-# action.beat.general.remove-unused-voice=
+# action.measure.general.remove-unused-voice=
 # action.beat.general.voice-up=
 # action.beat.general.voice-down=
 # action.beat.general.voice-auto=
