+++ common/resources/lang/messages_vi.properties
@@ -252,12 +252,12 @@ measure.clean=Xoá ô nhịp
 measure.remove=Bỏ ô nhịp
 measure.copy=Chép ô nhịp
 measure.paste=Dán ô nhịp
+# measure.voice.remove-unused=
 
 beat=Phách
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=Xoá nhịp
-# beat.voice.remove-unused=
 text.insert=Chèn văn bản
 text.editor=Soạn thảo văn bản
 text.text=Văn bản
@@ -642,7 +642,7 @@ action.measure.go-first=Tới ô nhịp đầu tiên
 action.measure.go-last=Tới ô nhịp cuối cùng
 action.measure.go-next=Tới ô nhịp kế
 action.measure.go-previous=Tới ô nhịp trước
-# action.beat.general.remove-unused-voice=
+# action.measure.general.remove-unused-voice=
 # action.beat.general.voice-up=
 # action.beat.general.voice-down=
 # action.beat.general.voice-auto=
