+++ common/resources/lang/messages_zh.properties
@@ -252,12 +252,12 @@ measure.clean=清除小节内容
 measure.remove=移除小节
 measure.copy=复制小节
 measure.paste=粘贴小节
+measure.voice.remove-unused=移除未使用交错拍子
 
 beat=编曲
 beat.insert-rest=插入休止符
 beat.delete-note-or-rest=删除休止符
 beat.clean=清除内容
-beat.voice.remove-unused=移除未使用交错拍子
 text.insert=插入文本
 text.editor=文本编辑器
 text.text=文本
@@ -642,7 +642,7 @@ action.measure.go-first=转到第一个小节
 action.measure.go-last=转到最后一个小节
 action.measure.go-next=转到下一个小节
 action.measure.go-previous=转到上一个小节
-action.beat.general.remove-unused-voice=移除未使用的交错拍子
+action.measure.general.remove-unused-voice=移除未使用的交错拍子
 action.beat.general.voice-up=符杆向上
 action.beat.general.voice-down=符杆向下
 action.beat.general.voice-auto=符杆方向(自动)
