+++ common/resources/lang/messages_zh_GB.properties
@@ -252,12 +252,12 @@ measure.clean=清除小节
 measure.remove=删除小节
 measure.copy=拷贝小节
 measure.paste=粘贴一小节
+measure.voice.remove-unused=删除未用交错拍子
 
 beat=音符
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=清除
-beat.voice.remove-unused=删除未用交错拍子
 text.insert=插入文本
 text.editor=文本编辑器
 text.text=文本
@@ -642,7 +642,7 @@ action.measure.go-first=第一小节
 action.measure.go-last=最后一小节
 action.measure.go-next=下一小节
 action.measure.go-previous=前一小节
-action.beat.general.remove-unused-voice=删除未用交错拍子
+action.measure.general.remove-unused-voice=删除未用交错拍子
 action.beat.general.voice-up=设置升高
 action.beat.general.voice-down=设置降低
 action.beat.general.voice-auto=自动
