+++ common/resources/lang/messages_zh_TW.properties
@@ -252,12 +252,12 @@ measure.clean=清除小節
 measure.remove=移除小節
 measure.copy=複製小節
 measure.paste=貼上小節
+measure.voice.remove-unused=移除沒用到的聲道
 
 beat=音符
 # beat.insert-rest=
 # beat.delete-note-or-rest=
 beat.clean=清除拍點
-beat.voice.remove-unused=移除沒用到的聲道
 text.insert=插入文字
 text.editor=文字編輯器
 text.text=文字
@@ -642,7 +642,7 @@ action.measure.go-first=到第一小節
 action.measure.go-last=到最後小節
 action.measure.go-next=到下一小節
 action.measure.go-previous=到上一小節
-action.beat.general.remove-unused-voice=移除沒用到的聲道
+action.measure.general.remove-unused-voice=移除沒用到的聲道
 action.beat.general.voice-up=設定上點(Set Stems Up)
 action.beat.general.voice-down=設定下點(Set Stems Down)
 action.beat.general.voice-auto=設定自動點(Set Auto Stems)
