+++ desktop/TuxGuitar/src/app/tuxguitar/app/action/installer/TGActionConfigMap.java
@@ -229,6 +229,7 @@ import app.tuxguitar.editor.action.measure.TGInsertMeasuresAction;
 import app.tuxguitar.editor.action.measure.TGPasteMeasureAction;
 import app.tuxguitar.editor.action.measure.TGRemoveMeasureAction;
 import app.tuxguitar.editor.action.measure.TGRemoveMeasureRangeAction;
+import app.tuxguitar.editor.action.measure.TGRemoveUnusedVoiceAction;
 import app.tuxguitar.editor.action.note.TGChangeNoteAction;
 import app.tuxguitar.editor.action.note.TGChangePickStrokeUpAction;
 import app.tuxguitar.editor.action.note.TGChangePickStrokeDownAction;
@@ -248,7 +249,6 @@ import app.tuxguitar.editor.action.note.TGMoveBeatsLeftAction;
 import app.tuxguitar.editor.action.note.TGMoveBeatsRightAction;
 import app.tuxguitar.editor.action.note.TGRemoveChordAction;
 import app.tuxguitar.editor.action.note.TGRemoveTextAction;
-import app.tuxguitar.editor.action.note.TGRemoveUnusedVoiceAction;
 import app.tuxguitar.editor.action.note.TGSetNoteFretNumberAction;
 import app.tuxguitar.editor.action.note.TGSetVoiceAutoAction;
 import app.tuxguitar.editor.action.note.TGSetVoiceDownAction;
