+++ desktop/TuxGuitar/src/app/tuxguitar/app/action/installer/TGActionInstaller.java
@@ -206,6 +206,7 @@ import app.tuxguitar.editor.action.measure.TGInsertMeasuresAction;
 import app.tuxguitar.editor.action.measure.TGPasteMeasureAction;
 import app.tuxguitar.editor.action.measure.TGRemoveMeasureAction;
 import app.tuxguitar.editor.action.measure.TGRemoveMeasureRangeAction;
+import app.tuxguitar.editor.action.measure.TGRemoveUnusedVoiceAction;
 import app.tuxguitar.editor.action.note.TGChangeNoteAction;
 import app.tuxguitar.editor.action.note.TGChangePickStrokeDownAction;
 import app.tuxguitar.editor.action.note.TGChangePickStrokeUpAction;
@@ -225,7 +226,6 @@ import app.tuxguitar.editor.action.note.TGMoveBeatsLeftAction;
 import app.tuxguitar.editor.action.note.TGMoveBeatsRightAction;
 import app.tuxguitar.editor.action.note.TGRemoveChordAction;
 import app.tuxguitar.editor.action.note.TGRemoveTextAction;
-import app.tuxguitar.editor.action.note.TGRemoveUnusedVoiceAction;
 import app.tuxguitar.editor.action.note.TGSetNoteFretNumberAction;
 import app.tuxguitar.editor.action.note.TGSetVoiceAutoAction;
 import app.tuxguitar.editor.action.note.TGSetVoiceDownAction;
